NetCDFProxy => ExperimentWithData (in progress)
