+++ atlas-data-storage/src/main/java/uk/ac/ebi/gxa/netcdf/ExperimentWithData.java
@@ -56,7 +56,8 @@ public class ExperimentWithData {
         return netCDFDao.getNetCDFDescriptors(experiment);
     }
 
-    private NetCDFProxy getProxy(ArrayDesign arrayDesign) throws AtlasDataException {
+    // TODO: change to private
+    public NetCDFProxy getProxy(ArrayDesign arrayDesign) throws AtlasDataException {
         NetCDFProxy p = proxies.get(arrayDesign);
         if (p == null) {
             p = netCDFDao.getNetCDFDescriptor(experiment, arrayDesign).createProxy();
